+++ test/NumSharp.UnitTest/Math/NDArray.Absolute.Test.cs
@@ -11,9 +11,9 @@ namespace NumSharp.UnitTest.Extensions
         [Test]
         public void absolute()
         {
-            //2D
+            //2D - np.abs now correctly preserves int dtype (NumPy-aligned)
             var n = np.arange(-2, 2).reshape(2, 2);
-            var n1 = np.abs(n).MakeGeneric<double>();
+            var n1 = np.abs(n).MakeGeneric<int>();
 
             Assert.IsTrue(n1[0, 0] == 2);
             Assert.IsTrue(n1[0, 1] == 1);
@@ -22,7 +22,7 @@ namespace NumSharp.UnitTest.Extensions
 
             //3D
             n = np.arange(-4, 4).reshape(2, 2, 2);
-            n1 = np.abs(n).MakeGeneric<double>();
+            n1 = np.abs(n).MakeGeneric<int>();
             Assert.IsTrue(n1[0, 0, 0] == 4);
             Assert.IsTrue(n1[0, 0, 1] == 3);
             Assert.IsTrue(n1[1, 0, 0] == 0);
@@ -30,7 +30,7 @@ namespace NumSharp.UnitTest.Extensions
 
             //4D
             n = np.arange(-12, 12).reshape(2, 3, 2, 2);
-            n1 = np.abs(n).MakeGeneric<double>();
+            n1 = np.abs(n).MakeGeneric<int>();
             Assert.IsTrue(n1[0, 0, 0, 0] == 12);
             Assert.IsTrue(n1[0, 1, 0, 0] == 8);
             Assert.IsTrue(n1[1, 2, 1, 1] == 11);
